pyproject.toml
[project]
name = "oms"
dynamic = [ "version" ]
description = "Sample Order Management System"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["fastapi (>=0.115.12,<0.116.0)", "uvicorn (>=0.34.0,<0.35.0)", "tortoise-orm (>=0.24.2,<0.25.0)", "pydantic (>=2.11.2,<3.0.0)", "openapi3 (>=1.8.2,<2.0.0)", "asyncpg (>=0.30.0,<0.31.0)", "pydantic-settings (>=2.8.1,<3.0.0)", "prometheus-fastapi-instrumentator (>=7.1.0,<8.0.0)", "celery (>=5.5.1,<6.0.0)", "redis (>=5.2.1,<6.0.0)", "aerich (>=0.8.2,<0.9.0)"]

[[project.authors]]
name = "Saeed Alizadeh"
email = "hellosaeedalizadeh@gmail.com"

[tool.poetry]
version = "0.1.0" # base version
packages = [
    { include = "app" },
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
mypy = "^1.15.0"
ruff = "^0.11.4"
httpx = "^0.28.1"
datamodel-code-generator = "^0.28.5"
tomlkit = "^0.13.2"
pytest-asyncio = "^0.26.0"
celery-stubs = "^0.1.3"

[tool.setuptools.packages.find]
include = ["app*"]
namespaces = false

[tool.mypy]
strict = true
files = "app"
namespace_packages = true
plugins = ["pydantic.mypy"]
exclude = ["app/api/v1/orders/schemas.py", "app/api/v1/analytics/schemas.py"]
ignore_missing_imports = true

[tool.ruff]
target-version = "py310"
line-length = 120
src = ["app", "tests"]
exclude = ["migrations", "tests/conftest.py", "tests/test_*.py", "app/api/v1/orders/schemas.py", "app/api/v1/analytics/schemas.py"]

[tool.ruff.lint]
extend-select = ["E", "W", "F", "I", "B", "C4", "UP", "ARG001", "FAST", "N", "Q", "T"]
ignore = ["B008", "B904"]

[tool.aerich]
tortoise_orm = "app.core.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "session"
